feat: get summary item by name

+++ internal/delivery/http/request/inventory_request.go
@@ -10,7 +10,7 @@ type CreateItemRequest struct {
 	ProductName string          `json:"productName" validate:"required"`
 	Status      entities.Status `json:"status" validate:"required,oneof=BUY SELL"`
 	Price       float64         `json:"price" validate:"required,min=0"`
-	Amount      int64           `json:"amount" validate:"required,min=0"`
+	Amount      int             `json:"amount" validate:"required,min=0"`
 	At          time.Time       `json:"at" validate:"required"`
 }
 
@@ -18,6 +18,6 @@ type UpdateItemRequest struct {
 	ProductName string          `json:"productName"`
 	Status      entities.Status `json:"status" validate:"oneof=BUY SELL"`
 	Price       float64         `json:"price" validate:"min=0"`
-	Amount      int64           `json:"amount" validate:"min=0"`
+	Amount      int             `json:"amount" validate:"min=0"`
 	At          time.Time       `json:"at"`
 }
